pyproject.toml
[project]
name = "dantec-graphql"
version = "0.0.0"
description = "Sets up and provides development tools for a GraphQL server."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "d>=0.2.2",
    "fastapi>=0.128.7",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "strawberry-graphql[cli,fastapi]>=0.291.3",
    "uvicorn[standard]>=0.40.0",
]
